requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
django-jazzmin==2.6.0
django-import-export==3.3.1
black==23.7.0